x86-64: guard exception stack dump against bad RSP, fix setexc() truncation
Addresses a further round of Copilot review feedback on PR #350:

- bios/arch/x86_64/panic.c, pgtable.c, pgtable.h: dump_stack()
  unconditionally dereferenced frame->rsp for its raw stack dump. That
  was fine while every exception meant a kernel-mode invariant had
  already broken so badly that nothing could be trusted; #350 makes a
  genuine ring-3 fault with an attacker- or bug-controlled (unmapped or
  non-canonical) user RSP an ordinary, caller-triggerable event instead.
  Faulting again mid-dump, still inside the original handler, risks
  losing the whole panic message to a recursive fault / #DF / triple
  fault. Added x86_64_addr_mapped_readable() (a read-only PML4/PDPT/PD
  walk via the physical-memory direct map, rejecting non-canonical
  addresses outright before they could alias a present page-table index)
  and guard both ends of the 16-qword dump with it. Verified with a
  throwaway test call against a known-mapped address, physical address 0
  (now backed by an allocated page per an earlier fix), an unmapped
  canonical address, and a non-canonical address -- all four came back
  correct -- then removed the test code.

- bios/bios.c, bios.h: setexc()'s vector parameter, its etv_* path's
  local, and its return type were all LONG (always 32-bit). The
  etv_timer/etv_critic/etv_term path (vecnum 0x100-0x102) round-trips a
  caller's vector through genuine C function pointers -- 64-bit ones on
  this arch -- directly, so a LONG parameter/return silently truncated a
  higher-half caller's vector before it ever reached etv_timer's own
  assignment; this arch's own bdosmain.c calls exactly that path today
  (Setexc(0x21, (long)enter), Setexc(0x22, (long)bdos_trap2)). Widened
  to native `long` for that path (a no-op on m68k/ARM, where long and
  LONG are the same width). The separate `addr = (LONG *)(4L * num)`
  path (every other vecnum) is deliberately NOT widened: that low-memory
  table is fundamentally 32-bit-per-slot by historical m68k ABI
  convention, unrelated to this arch's own pointer width, and can't
  round-trip a genuine 64-bit pointer regardless of this function's own
  C-level parameter type -- noted in a comment cross-referencing #351,
  which already tracks the same "kernel pointer needs to fit in a
  narrower ABI slot" theme.

Rebuilt and boot-tested under QEMU/OVMF (pc-x86_64): reaches the same
gouser() milestone as before; the raw stack dump still prints correctly
for the legitimate kernel-mode panic path, confirming the new guard
doesn't spuriously suppress it. Compile-tested virt-arm (bios.c/bios.h
are shared).

+++ bios/bios.c
@@ -1346,9 +1346,30 @@ static LONG bios_4(WORD r_w, UBYTE *adr, WORD numb, WORD first, WORD drive, LONG
  *
  */
 
-LONG setexc(WORD num, LONG vector)
+/*
+ * `long`, not portab.h's always-32-bit LONG, for the parameter, the
+ * etv_*-path local, and the return type: etv_timer/etv_critic/etv_term
+ * are genuine C function pointers (real 64-bit ones on x86-64's LP64),
+ * and num values 0x100-0x102 round-trip a caller's vector through them
+ * directly (see that path's own comment below) rather than through the
+ * fixed-32-bit-per-slot low-memory table the num=0x21/0x22 (etc.) path
+ * below still uses -- a LONG parameter/return here silently truncated
+ * any higher-half caller (e.g. bdosmain.c's own
+ * Setexc(0x21, (long)enter)) before it ever reached etv_timer's own
+ * assignment. A no-op on m68k/ARM, where long and LONG are the same
+ * width.
+ *
+ * The `addr = (LONG *)(4L * num)` path below is NOT similarly widened,
+ * deliberately: that low-memory table is fundamentally 32-bit-per-slot
+ * by historical (m68k) ABI convention, unrelated to this arch's own
+ * pointer width, and widening the slot itself would be a much larger
+ * change than this function's own parameter type. See kelihlodversson/
+ * pTOS#351 for the same "kernel pointer needs to fit in a narrower ABI
+ * slot" theme applied to that path specifically.
+ */
+long setexc(WORD num, long vector)
 {
-    LONG oldvector;
+    long oldvector;
     LONG *addr;
 
     /*
@@ -1368,17 +1389,17 @@ LONG setexc(WORD num, LONG vector)
     switch (num)
     {
     case 0x100:
-        oldvector = (LONG)etv_timer;
+        oldvector = (long)etv_timer;
         if (vector != -1)
             etv_timer = (void(*)(int))vector;
         return oldvector;
     case 0x101:
-        oldvector = (LONG)etv_critic;
+        oldvector = (long)etv_critic;
         if (vector != -1)
             etv_critic = (LONG(*)(WORD,WORD))vector;
         return oldvector;
     case 0x102:
-        oldvector = (LONG)etv_term;
+        oldvector = (long)etv_term;
         if (vector != -1)
             etv_term = (void(*)(void))vector;
         return oldvector;
@@ -1388,7 +1409,7 @@ LONG setexc(WORD num, LONG vector)
     oldvector = *addr;
 
     if(vector != -1) {
-        *addr = vector;
+        *addr = (LONG)vector;
     }
     return oldvector;
 }
